move to distributed.py and test_distributed.py
